Made overall tests more stable
Ethereum has a niptick around gas consumption and zero bytes. Indeed, the zero bytes within an input are costing less than non-zero bytes. This is usually not a problem but when we deal with random addresses and contract deployment, it can be a problem for stability of tests.

To overcome the issue, we are changing a bit how generate random inputs and deploy contracts:
- Contract deployment now ensures the deployed contract don't contains any zeros, re-do another deployment if it didn't work
- Random hex we uses to generate now never generates 0 anymore.